requirements.txt
aiogram>=3.0
pandas
openpyxl
python-dotenv
gspread
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
